requirements.txt
fastapi
uvicorn
runpod
python-dotenv
asyncio
aiohttp
websocket_client
pillow
numpy